FIX : only active contest submission
